Add not-darwin code
